* Refactored a bunch of things to support latest version of Discord Akairo framework.
* Updated most of the other dependencies.

* Various other things.

+++ src/struct/LCommandHandler.js
@@ -1,14 +1,16 @@
 const { CommandHandler } = require('discord-akairo')
 
 class LCommandHandler extends CommandHandler {
-  constructor (client, options = {}) {
+  constructor (client, options) {
     super(client, options)
 
     const {
-      statusTimeout = -1
-    } = client.akairoOptions
+      statusTimeout = -1,
+      purgeCommandsTimeout = -1
+    } = options
 
     this.statusTimeout = statusTimeout
+    this.purgeCommandsTimeout = purgeCommandsTimeout
 
     this.statusTemplates = {
       success: {
@@ -53,45 +55,36 @@ class LCommandHandler extends CommandHandler {
   async handle (message) {
     if (message.author.id === this.client.user.id) {
       this.client.stats.increment('messages-sent')
-
-      if (!message.status) {
-        message.status = this._status
-      }
+      if (!message.status) { message.status = this._status }
     } else {
       this.client.stats.increment('messages-received')
     }
-
     return super.handle(message)
   }
 
   load (thing, isReload) {
     const mod = super.load(thing, isReload)
-
-    if (this.client.stats.get('initiated') && isReload && mod && typeof mod.onReady === 'function') {
-      mod.onReady()
-    }
-
+    if (this.ready && isReload && mod && typeof mod.onReady === 'function') { mod.onReady() }
     return mod
   }
 
   reload (id) {
     const mod = this.modules.get(id.toString())
-
-    if (mod && typeof mod.onReload === 'function') {
-      mod.onReload()
-    }
-
+    if (mod && typeof mod.onReload === 'function') { mod.onReload() }
     return super.reload(id)
   }
 
   remove (id) {
     const mod = this.modules.get(id.toString())
-
-    if (mod && typeof mod.onRemove === 'function') {
-      mod.onRemove()
-    }
-
+    if (mod && typeof mod.onRemove === 'function') { mod.onRemove() }
     return super.remove(id)
+  }
+
+  readyAll () {
+    this.ready = true
+    this.modules.forEach(mod => {
+      if (typeof mod.onReady === 'function') { mod.onReady() }
+    })
   }
 }
